requirements.txt
casadi==3.5.1
certifi==2020.4.5.1
cycler==0.10.0
entrypoints==0.3
flake8==3.7.9
kiwisolver==1.2.0
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.3
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0
svg.path==4.0.2
